requirements.txt
Streamlit==1.10.0
requests==2.26.0
python-dotenv==0.19.2
openai==0.27.0
llama==0.1.0
